consolidate websocket tests
